small refactor using BigInteger with describe() method. removed testReserveIdentifier test since the reserveIdentifier method is no longer in the service api. updating pom to 0.6.4-SNAPSHOT, and commenting out test-class move command used for packaging webtester (since it breaks normal use)
git-svn-id: file:///var/code/dataone/dataone/software/cicore/trunk/d1_integration@4659 a8e578cc-b067-0410-9529-a28e5a6d5a41

+++ src/test/java/org/dataone/integration/it/ExampleUtilities.java
@@ -25,6 +25,7 @@ import static org.junit.Assert.assertThat;
 
 import java.io.IOException;
 import java.io.InputStream;
+import java.math.BigInteger;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.util.ArrayList;
@@ -397,7 +398,7 @@ public class ExampleUtilities {
         SystemMetadata sysmeta = new SystemMetadata();
         sysmeta.setIdentifier(guid);
         sysmeta.setObjectFormat(objectFormat);
-        sysmeta.setSize(12);
+        sysmeta.setSize(BigInteger.valueOf(12));
         Subject submitter = new Subject();
         String dn = "uid=jones,o=NCEAS,dc=ecoinformatics,dc=org";
         submitter.setValue(dn);
